Removed OrderCustomerActionListener and added OrderCustomerHelper
